pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "machinations"
version = "1.0.0"
description = "A Game of AI Diplomacy & Betrayal — pit frontier LLMs against each other in a strategy game"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MACHINATIONS Contributors"},
]
keywords = ["ai", "llm", "game", "diplomacy", "strategy", "agents", "deepseek", "claude", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment :: Turn Based Strategy",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "rich>=13.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/machinations"
Repository = "https://github.com/yourusername/machinations"

[project.scripts]
machinations = "run:main"

[tool.setuptools.packages.find]
include = ["machinations*"]
